+++ Assets/Scripts/Enums/TypeCommunication.cs
@@ -1,5 +1,5 @@
 /// <summary>対話における思考タイプ定義。</summary>
-public enum TypeCommunication : byte
+public enum TypeCommunication : int
 {
     /// <summary>明確化指向。</summary>
     /// <remarks>結論ありきで詳細を詰めていくタイプ。</remarks>
