+++ Assets/Scripts/Enums/TypeResponse.cs
@@ -1,5 +1,5 @@
 /// <summary>素質を持つ、立ち位置タイプ定義。</summary>
-public enum TypeResponse : byte
+public enum TypeResponse : int
 {
     /// <summary>現場タイプ。</summary>
     /// <remark>知らない人々の中でも、比較的活動できるタイプ。</remark>
